requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
faiss-cpu==1.12.0
fastapi==0.118.0
filetype==1.2.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-nvidia-ai-endpoints==0.3.18
langchain-text-splitters==0.3.11
langsmith==0.4.31
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
propcache==0.3.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.20.1
zstandard==0.25.0
